feat: encapsulate circle geometry
